# src/main/java/org/kiss/model/entity/util/CompanyAddress.java
package org.kiss.model.entity.util;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class CompanyAddress {

    private String street;
    private String city;
    private String postalCode;
    private String country;
}