# src/main/java/org/kiss/model/entity/util/CoveragePoint.java
package org.kiss.model.entity.util;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class CoveragePoint {

    private String label;
    private boolean covered;
}