# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Car Rental API
  description: Backend API for the car rental application.
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/companies:
    get:
      tags:
        - Companies
      operationId: listCompanies
      summary: List car rental companies as a lightweight overview. Restricted to admin accounts.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: A page of company overviews.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyPage"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Companies
      operationId: createCompany
      summary: Create a car rental company. Restricted to admin accounts.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyRegistrationRequest"
      responses:
        "201":
          description: Company created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "400":
          description: Validation error, or one of the user ids does not reference an existing user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A company with this registration number is already registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/companies/{id}:
    get:
      tags:
        - Companies
      operationId: getCompanyById
      summary: Get the full details of a single company. Restricted to admin accounts.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No company exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - Companies
      operationId: updateCompany
      summary: Partially update a company. Restricted to admin accounts. Only fields present in the request body are changed.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdateRequest"
      responses:
        "200":
          description: Company updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "400":
          description: Validation error, or one of the user ids does not reference an existing user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No company exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A company with this registration number is already registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/companies/{id}/users/{userId}:
    put:
      tags:
        - Companies
      operationId: addUserToCompany
      summary: >
        Add an existing user to a company. Restricted to admin accounts. Idempotent — adding a user
        already in the company is a no-op. If the user currently belongs to a different company, they
        are moved (removed from the old company's user list).
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The updated company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "400":
          description: The user is an admin account and cannot be attached to a company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No company or no user exists with the given id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Companies
      operationId: removeUserFromCompany
      summary: >
        Remove a user from a company. Restricted to admin accounts. Idempotent — removing a user who
        isn't currently a member of the company is a no-op.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The updated company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No company or no user exists with the given id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/insurances:
    get:
      tags:
        - Insurances
      operationId: listInsurances
      summary: List all insurance tiers. Public.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: A page of insurance tiers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsurancePage"
    post:
      tags:
        - Insurances
      operationId: createInsurance
      summary: Create an insurance tier. Restricted to admin accounts.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsuranceRequest"
      responses:
        "201":
          description: Insurance created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsuranceResponse"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: An insurance with this name already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/insurances/{id}:
    get:
      tags:
        - Insurances
      operationId: getInsuranceById
      summary: Get the details of an insurance tier. Public.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The insurance tier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsuranceResponse"
        "404":
          description: No insurance exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Insurances
      operationId: updateInsurance
      summary: Update an insurance tier. Restricted to admin accounts.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsuranceRequest"
      responses:
        "200":
          description: Insurance updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsuranceResponse"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No insurance exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Another insurance with this name already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Insurances
      operationId: deleteInsurance
      summary: Delete an insurance tier. Restricted to admin accounts.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Insurance deleted.
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authenticated but not an admin account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No insurance exists with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRole:
      type: string
      description: All roles a user account can have.
      enum:
        - USER
        - DEALERSHIP
        - ADMIN
...
    UserProfileResponse:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - role
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        companyId:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - status
        - message
        - timestamp
      properties:
        status:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    CompanyStatus:
      type: string
      description: Verification status, controlled by admins.
      enum:
        - PENDING_VERIFICATION
        - ACTIVE
        - SUSPENDED
    CompanyAddress:
      type: object
      required:
        - street
        - city
        - country
      properties:
        street:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        postalCode:
          type: string
        country:
          type: string
          minLength: 1
    CompanyRegistrationRequest:
      type: object
      required:
        - name
      properties:
        userIds:
          type: array
          items:
            type: string
          description: User ids attached to this company. May be omitted, null, or empty.
        name:
          type: string
          minLength: 1
        registrationNumber:
          type: string
          maxLength: 50
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          maxLength: 30
        address:
          $ref: "#/components/schemas/CompanyAddress"
        website:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        logoUrl:
          type: string
    CompanyUpdateRequest:
      type: object
      description: Partial update — only properties present in the request body are changed.
      properties:
        userIds:
          type: array
          items:
            type: string
          description: Replaces the current user list when present.
        name:
          type: string
          minLength: 1
        registrationNumber:
          type: string
          maxLength: 50
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          maxLength: 30
        address:
          $ref: "#/components/schemas/CompanyAddress"
        website:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        logoUrl:
          type: string
        status:
          $ref: "#/components/schemas/CompanyStatus"
    CompanyResponse:
      type: object
      required:
        - id
        - users
        - name
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfileResponse"
          description: Full profile of every user attached to this company.
        name:
          type: string
        registrationNumber:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        address:
          $ref: "#/components/schemas/CompanyAddress"
        website:
          type: string
        description:
          type: string
        logoUrl:
          type: string
        status:
          $ref: "#/components/schemas/CompanyStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CompanySummaryResponse:
      type: object
      required:
        - id
        - name
        - status
        - createdAt
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string
        country:
          type: string
        logoUrl:
          type: string
        status:
          $ref: "#/components/schemas/CompanyStatus"
        createdAt:
          type: string
          format: date-time
    CompanyPage:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CompanySummaryResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...
    CoveragePoint:
      type: object
      required:
        - label
        - covered
      properties:
        label:
          type: string
          minLength: 1
        covered:
          type: boolean
    InsuranceRequest:
      type: object
      required:
        - name
        - dailyPriceAmount
      properties:
        name:
          type: string
          minLength: 1
        coveragePoints:
          type: array
          items:
            $ref: "#/components/schemas/CoveragePoint"
        dailyPriceAmount:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
        dailyPriceCurrency:
          type: string
          default: EUR
        hasFranchise:
          type: boolean
          default: false
        franchiseAmount:
          type: number
          format: double
          minimum: 0
    InsuranceResponse:
      type: object
      required:
        - id
        - name
        - dailyPriceAmount
        - dailyPriceCurrency
        - hasFranchise
      properties:
        id:
          type: string
        name:
          type: string
        coveragePoints:
          type: array
          items:
            $ref: "#/components/schemas/CoveragePoint"
        dailyPriceAmount:
          type: number
          format: double
        dailyPriceCurrency:
          type: string
        hasFranchise:
          type: boolean
        franchiseAmount:
          type: number
          format: double
    InsurancePage:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/InsuranceResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer